requirements.txt
streamlit
langchain>=0.3.0
langchain-google-genai
langchain-community
langchain-core
chromadb
python-dotenv